Structure repository for implementation upload
Because the implementation submission for Part III explicitly does not
require prior information from the methodology, background research, et
cetera (these will only be included in the final report), it was
necessary to divide the latex files in the `latex/` directory into two
seperate documents: one for the project proposal and methodology (from
Part I and Part II), and a second document (currently empty) for the
implementation itself.

As the implementation submission also requires the working project code,
I have also structured the document to have a `python/` and `datasets/`
directory.