Add initial evaluation models to python/
This commit completes the first part of the implementation upload. Here
we add three jupyter notebooks which are used to train the initial
models that are used as evaluation baselines. These models are
implemented using a mix of functionality held in the `python/common/`
directory, which is a package that is used to implement utility
functions that are shared between all notebooks, as well as some code
held within the notebooks themselves.

The code is designed to be ran either on a local machine with a powerful
GPU, or via a remote jupyter server kernel from a compute provider.