Update dataset/ README.md
The README.md file in the `dataset/` directory is updated to include
further information on how the dataset is split for training,
validation, and testing, as well as additional information on how the
labels and features are encoded.

The actual dataset is still not cleared for release at the moment. Once
the dataset itself is available, additional information will be added.