Add lab notebooks (.ipynb) to python/
The raw data collected from the hyperparameter search process is saved
to disk as Python pickle files. These files are Python objects (lists of
dictionaries) containing the raw data collected from hyperparameter
search and model k-fold cross-validation. This data has to be
intepreted, and processed in order to yield insights.

The two lab notebooks included in this commit are
regime-1-analysis.ipynb and regime-2-analysis.ipynb. They are used to
create the processed data files which are later graphed in the LaTeX
repository.

Likewise the raw data files are also included in this commit. This makes
the analysis repeatable by other readers.