Add labels and tf.data.Dataset build .ipynb
The process to create a `tf.data.Dataset` object is non-intuitive. In
this commit, we upload the Jupyter notebook that is used to create these
dataset objects from the raw labels and image files. Additionally, we
also upload the dataset labels as a one-hot encoded csv and Python
pickle object.