Complete section on k-fold cross-validation
In this section on the implementation, I talk about the k-fold
cross-validation implementation. I try to keep this section relatively
short, since it is only secondary to the actual running of the models.

Additionally this commit corrects an error in the
`latex/graphs/hypersearch-regime-2.tex` graph by explicitly adding the
number of rows into the TiKZ/PGFPlots mesh plot call.