Add training & analysis `.ipynb` for final model
This commit adds the training and analysis `.ipynb` for the final model,
which is the iteration of the model trained using the optimal
hyperparameters found during hyperparameter search. The directory
located in `python/final-model/` contains the jupyter notebook files
used to train the model, as well as the Keras checkpoint directory used
to load the model, and also processed data `.csv` files.

Future commits will include `.ipynb` files that will demonstrate
evaluation and inference.

Note: the keras checkpoint directory has an uncompressed filesize of 127
megabytes, and when compressed using `xz` with maximal compression
options (`--OPT=9e`) yields a compressed size of 110 megabytes. This is
over the Github threshold limit for large files, which is 100mb. Hence,
the file itself is stored using `git-lfs`.