Update `dataset/` for git-lfs support
This commit compresses the `tf.Data.dataset` directory objects using
`xz` in preparation for their eventual upload and storage with `git-lfs`
once we obtain permission to release the data.

The changes mostly involve updates to the documentation, and the
corresponding `.gitignore` file as the dataset is not yet tracked on
git.