Create `latex/standalone/` for standalone figures
As we gear up for the poster presentation, there are some LaTeX graphs
and figures from the original project report that I would like to export
individually for use in the presentation and external files. In order to
do so, we use the LaTeX `standalone` package alongside with this new
directory containing standalone files.

These files are essentially copies of figures from `latex/graphs/` or
`latex/figures/`, modified to generate PDF files which can be easily
embedded elsewhere. Modifications include changing the size of the
graphs, or font.