Refactor latex layout in prep. for final report
This commit significantly changes the internal layout of the `latex/`
directory, in order to prepare the document for submission as a final
report. We create a `latex/frontmatter/` and `latex/backmatter/`
directory to hold the front and back-matter, respectively. We also
create additional files to hold the declaration of authorship, abstract,
and acknowledgements. Finally, a new root file `final-report.tex` is
created to hold everything together.